BufferProvider: integrate and verify heuristic trigger of clean-up
The mechanism is tuned now to engage rather quickly
whenever excess capacity builds up in a local allocator pool,
but when all buffers are used most of the time, they will be retained.

Can now demonstrate by integration test that capacity is sent back.

+++ src/vault/mem/local-mem-pool.hpp
@@ -205,9 +205,9 @@ namespace mem   {
       
       /** turnover-based trigger for heuristic pool clean-up */
       bool
-      shouldCleanup (float sharpen =1.0f)  const
+      shouldCleanup (double sharpen =1.0f)  const
         {
-          return turnover_ * float(MATCH_SCORE) <= maxScore_ * sharpen;
+          return turnover_ * double(MATCH_SCORE) <= maxScore_ * sharpen;
         }
       
       /**
